Use seal-trait to put control  in the upstream .
