+++ src/lib.rs
@@ -938,61 +938,3 @@ impl Command {
         async { child?.output().await }
     }
 }
-
-#[cfg(unix)]
-/// Moves `Fd` out of nonblocking mode.
-pub fn set_blocking<T: std::os::unix::io::AsRawFd>(stdout: T) -> io::Result<T> {
-    unsafe {
-        let res = libc::fcntl(stdout.as_raw_fd(), libc::F_GETFL);
-        let errno = libc::fcntl(
-            stdout.as_raw_fd(),
-            libc::F_SETFL,
-            !(!res | libc::O_NONBLOCK),
-        );
-
-        // Unix-like systems when errno is_minus_one then return last_os_error.
-        if errno == -1 {
-            return Err(io::Error::last_os_error());
-        }
-    }
-
-    Ok(stdout)
-}
-
-mod test {
-
-    #[cfg(unix)]
-    #[test]
-    fn test_into_inner() {
-        futures_lite::future::block_on(async {
-            use crate::{set_blocking, Command};
-
-            use std::io::{Read, Result};
-            use std::process::Stdio;
-            use std::str::from_utf8;
-
-            let mut ls_child = Command::new("cat")
-                .arg("Cargo.toml")
-                .stdout(Stdio::piped())
-                .spawn()?;
-
-            let stdio: Stdio = ls_child.stdout.take().unwrap().into_inner().await?.into();
-
-            let mut echo_child = Command::new("grep")
-                .arg("async")
-                .stdin(stdio)
-                .stdout(Stdio::piped())
-                .spawn()?;
-
-            let mut buf = vec![];
-            let mut stdout = echo_child.stdout.take().unwrap().into_inner().await?;
-            stdout = set_blocking(stdout)?;
-
-            stdout.read_to_end(&mut buf)?;
-            dbg!(from_utf8(&buf).unwrap_or(""));
-
-            Result::Ok(())
-        })
-        .unwrap();
-    }
-}
